# web-morda/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cookbook",
    "version": "1.0"
  },
  "paths": {
...
    "/api/products": {
      "post": {
        "tags": [
          "Products"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Photos": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "PhotoFiles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Calories": {
                    "type": "number",
                    "format": "float"
                  },
                  "Proteins": {
                    "type": "number",
                    "format": "float"
                  },
                  "Fats": {
                    "type": "number",
                    "format": "float"
                  },
                  "Carbs": {
                    "type": "number",
                    "format": "float"
                  },
                  "Composition": {
                    "type": "string"
                  },
                  "Category": {
                    "$ref": "#/components/schemas/ProductCategory"
                  },
                  "CookingType": {
                    "$ref": "#/components/schemas/CookingType"
                  },
                  "Flags": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DietaryFlag"
                    }
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Photos": {
                  "style": "form"
                },
                "PhotoFiles": {
                  "style": "form"
                },
                "Calories": {
                  "style": "form"
                },
                "Proteins": {
                  "style": "form"
                },
                "Fats": {
                  "style": "form"
                },
                "Carbs": {
                  "style": "form"
                },
                "Composition": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "CookingType": {
                  "style": "form"
                },
                "Flags": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "Category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductCategory"
            }
          },
          {
            "name": "CookingType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CookingType"
            }
          },
          {
            "name": "Vegan",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "GlutenFree",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "SugarFree",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductSortField"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Photos": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "PhotoFiles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Calories": {
                    "type": "number",
                    "format": "float"
                  },
                  "Proteins": {
                    "type": "number",
                    "format": "float"
                  },
                  "Fats": {
                    "type": "number",
                    "format": "float"
                  },
                  "Carbs": {
                    "type": "number",
                    "format": "float"
                  },
                  "Composition": {
                    "type": "string"
                  },
                  "Category": {
                    "$ref": "#/components/schemas/ProductCategory"
                  },
                  "CookingType": {
                    "$ref": "#/components/schemas/CookingType"
                  },
                  "Flags": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DietaryFlag"
                    }
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Photos": {
                  "style": "form"
                },
                "PhotoFiles": {
                  "style": "form"
                },
                "Calories": {
                  "style": "form"
                },
                "Proteins": {
                  "style": "form"
                },
                "Fats": {
                  "style": "form"
                },
                "Carbs": {
                  "style": "form"
                },
                "Composition": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "CookingType": {
                  "style": "form"
                },
                "Flags": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CookingType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "DietaryFlag": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "FlagFilterOption": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProductCategory": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "type": "number",
            "format": "float"
          },
          "proteins": {
            "type": "number",
            "format": "float"
          },
          "fats": {
            "type": "number",
            "format": "float"
          },
          "carbs": {
            "type": "number",
            "format": "float"
          },
          "composition": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategory"
          },
          "cookingType": {
            "$ref": "#/components/schemas/CookingType"
          },
          "flags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DietaryFlag"
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductSortField": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "SortDirection": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  },
...
}